% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{org.Hbacteriophora.eg}
\alias{org.Hbacteriophora.eg}
\title{Access Heterorhabditis bacteriophora annotation data}
\usage{
org.Hbacteriophora.eg()
}
\description{
Access Heterorhabditis bacteriophora annotation data
}
